requirements.txt
requests
pulsar-client
pymongo
flask
plotly
